refactor: 불명확한 함수 제거, types 디렉토리 생성
