+++ frontend/src/types/index.ts
@@ -1,0 +1,12 @@
+type Dictionary = { [key: string]: unknown };
+
+type JSONType =
+  | string
+  | number
+  | boolean
+  | null
+  | undefined
+  | JSONType[]
+  | { [key: string]: JSONType };
+
+export type { Dictionary, JSONType };
